requirements.txt
python-telegram-bot==20.5
python-decouple==3.8